move input features to their own files
